Cucumber tester for beregning av barnetilsyn med flere perioder
